Refactor main.rs by extracting logic into separate modules
- Created `file_utils.rs` for file-related functions.
- Created `csv_operations.rs` for CSV-related functions.
- Created `json_operations.rs` for JSON-related functions.
- Updated `main.rs` to use these new modules.
- Improved code organization and maintainability.
